Use underscore to mark out private methods

+++ downloaders/pdf_downloader.py
@@ -22,11 +22,11 @@ class PDFDownloader:
     def __init__(self, is_headless: bool = False):
         self.is_headless = is_headless
         self.directory = Directory(self.is_headless)
-        self.driver = self.initialize_driver()
+        self.driver = self._initialize_driver()
         self.wait_time = WaitTime()
 
     # Methods for managing PDFDownloader's life cycle
-    def initialize_driver(self):
+    def _initialize_driver(self):
         chrome_options = webdriver.ChromeOptions()
         chrome_options.add_argument('--incognito')
         chrome_options.add_argument('--window-size=1920,1080')
@@ -60,23 +60,23 @@ class PDFDownloader:
         self.driver.quit()
 
     # Methods for generating PDF
-    def save_current_page_as_pdf_in_output_folder(self, output_path_with_filename: str):
+    def _save_current_page_as_pdf_in_output_folder(self, output_path_with_filename: str):
         if self.is_headless:
-            self.write_to_local_file_in_output_folder(output_path_with_filename)
+            self._write_to_local_file_in_output_folder(output_path_with_filename)
         else:
-            self.write_to_temp_folder_and_move_to_output_folder(output_path_with_filename)
+            self._write_to_temp_folder_and_move_to_output_folder(output_path_with_filename)
 
-    def write_to_local_file_in_output_folder(self, output_path_with_filename: str):
+    def _write_to_local_file_in_output_folder(self, output_path_with_filename: str):
         b64_data = self.driver.execute_cdp_cmd("Page.printToPDF", {
             "printBackground": True,
         })['data']
         b64_data_decoded = b64decode(b64_data, validate=True)
-        PDFDownloader.validate_b64_string(b64_data_decoded)
+        PDFDownloader.validate_b64_string_is_pdf(b64_data_decoded)
         with open(output_path_with_filename, "wb") as f:
             f.write(b64_data_decoded)
 
     # Methods to do with waiting
-    def write_to_temp_folder_and_move_to_output_folder(self, filename_path_output: str):
+    def _write_to_temp_folder_and_move_to_output_folder(self, filename_path_output: str):
         self.driver.execute_script('window.print();')  # download PDF to temp directory
         # move file from temp directory to output directory and rename file
         for _, _, filenames in os.walk(self.directory.temp_path):
@@ -86,7 +86,7 @@ class PDFDownloader:
                     filename_path_temp = os.path.join(self.directory.temp_path, filename_temp)
                     os.rename(filename_path_temp, filename_path_output)
 
-    def wait_for_element_to_load(self, by: By, element_target: str) -> bool:
+    def _wait_for_element_to_load(self, by: By, element_target: str) -> bool:
         try:
             WebDriverWait(self.driver, self.wait_time.max_wait_time).until(
                 EC.presence_of_element_located((by, element_target)))
@@ -101,7 +101,7 @@ class PDFDownloader:
     #  Tried and failed:
     #  WebDriverWait(driver, 10).until(lambda driver: driver.execute_script('return document.readyState') == 'complete')
     #  WebDriverWait(driver, 10).until(lambda driver: driver.execute_script('return "return jQuery.active') == '0')
-    def wait_for_page_to_finish_loading(self):  # TODO WIP
+    def _wait_for_page_to_finish_loading(self):  # TODO WIP
         try:
             WebDriverWait(self.driver, self.wait_time.max_wait_time).until(
                 staleness_of(self.driver.find_element_by_tag_name('html')))
@@ -123,25 +123,28 @@ class Directory:
         self.path_to_directory = os.path.dirname(__file__)
         self.chromedriver_path = os.path.join(self.path_to_directory, '../utilities/chromedriver')
         self.output_path = os.path.join(self.path_to_directory, '../output')
-        self.raise_error_if_chromedriver_missing()
-        self.ensure_output_folder_exists()
+        self._raise_error_if_chromedriver_missing()
+        self._ensure_output_folder_exists()
         if not is_headless:
             self.temp_path = os.path.join(self.path_to_directory, 'temp')
-            self.ensure_temp_folder_exists()
+            self._ensure_temp_folder_exists()
 
-    def raise_error_if_chromedriver_missing(self):
+    # Methods used in initialization
+    def _raise_error_if_chromedriver_missing(self):
         if not os.path.exists(self.chromedriver_path):
             raise errors.ChromedriverMissing(f"chromedriver does not exist at {self.chromedriver_path}")
 
-    def ensure_temp_folder_exists(self):
+    def _ensure_temp_folder_exists(self):
         Directory.ensure_folder_exists(self.temp_path)
 
-    def ensure_output_folder_exists(self):
+    def _ensure_output_folder_exists(self):
         Directory.ensure_folder_exists(self.output_path)
 
+    # Tell Directory object to delete its temp folder (as part of wrapping up the program)
     def delete_temp_folder(self):
-        os.rmdir(self.temp_path)  # remove empty directory
-        # shutil.rmtree(self.temp_path)  # delete directory and all its contents
+        if os.path.isdir(self.temp_path):
+            os.rmdir(self.temp_path)  # remove empty directory
+            # shutil.rmtree(self.temp_path)  # delete directory and all its contents
 
     @staticmethod
     def ensure_folder_exists(path_to_folder: str):
